feat:1.增加链表clist.h fix:1.修改log的 一处bug，导致发送的数据长度不正确。

+++ utilities/log/log.c
@@ -204,7 +204,7 @@ size_t log_strcpy(size_t cur_len, char *dst, const char *src)
 void log_output(uint8_t level, const char *tag, const char *file, const char *func, const long line, const char *format, ...)
 {
     log_output_lock();
-    size_t log_len = 0, fmt_result = 0, newline_len = strlen(LOG_NEWLINE_SIGN);
+    int log_len = 0, fmt_result = 0, newline_len = strlen(LOG_NEWLINE_SIGN);
     char line_num[LOG_LINE_NUM_MAX_LEN + 1] = {0};
 
     va_list args;
